Testes de produtos

+++ funcoes/produtos.py
@@ -5,7 +5,7 @@ def getProduto(id):
     query = "SELECT * FROM produto WHERE id = %s;"
     parametros = [id]
     dadosProduto = baseDelivery.selecionarUm(query, parametros)
-    return dadosProduto
+    return dadosProduto if dadosProduto else None
 
 
 def getListaProdutos():
@@ -16,31 +16,34 @@ def getListaProdutos():
 
 @chaveNecessaria
 def criarProduto(nome, descricao, valor, imagem):
-    query = "INSERT INTO produto (nome, descricao, valor, imagem) VALUES (%s, %s, %s, %s) RETURNING id;"
-    parametros = [nome, descricao, valor, imagem]
-    dados = baseDelivery.executar(query, parametros)
-    id = dados['id']
+    if valor > 0:
+        query = "INSERT INTO produto (nome, descricao, valor, imagem) VALUES (%s, %s, %s, %s) RETURNING id;"
+        parametros = [nome, descricao, valor, imagem]
+        dados = baseDelivery.executar(query, parametros)
+        id = dados['id']
 
-    dadosProduto = getProduto(id)
-    return dadosProduto
+        dadosProduto = getProduto(id)
+        return dadosProduto
 
 
 @chaveNecessaria
 def atualizarProduto(nome, descricao, valor, imagem, id):
-    query = "UPDATE produto SET nome = %s, descricao = %s, valor = %s, imagem = %s WHERE id = %s;"
-    parametros = [nome, descricao, valor, imagem, id]
-    baseDelivery.executar(query, parametros)
+    if valor > 0:
+        query = "UPDATE produto SET nome = %s, descricao = %s, valor = %s, imagem = %s WHERE id = %s;"
+        parametros = [nome, descricao, valor, imagem, id]
+        baseDelivery.executar(query, parametros)
 
-    dadosProduto = getProduto(id)
-    return dadosProduto
+        dadosProduto = getProduto(id)
+        return dadosProduto
 
 
 @chaveNecessaria
 def deletarProduto(id):
-    query = "DELETE FROM produto WHERE id = %s;"
-    parametros = [id]
-    baseDelivery.executar(query, parametros)
+    if getProduto(id):
+        query = "DELETE FROM produto WHERE id = %s;"
+        parametros = [id]
+        baseDelivery.executar(query, parametros)
 
-    dadosProduto = getProduto(id)
-    foiDeletado = not dadosProduto
-    return foiDeletado
+        dadosProduto = getProduto(id)
+        foiDeletado = not dadosProduto
+        return foiDeletado
